refactor: move sort map objects by key method to the utils
