+++ lib/layout/data_grid.dart
@@ -15,6 +15,7 @@ import 'package:ensemble/widget/widget_util.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
+import 'package:ensemble/framework/view/page.dart';
 
 enum DataColumnSortType { ascending, descending }
 
@@ -459,54 +460,14 @@ class DataGridState extends WidgetState<DataGrid>
 
       if (sortable != null && sortable) {
         bool isAscendingOrder = sortOrder == DataColumnSortType.ascending.name;
-        dataList = sortMapObjectsByKey(dataList, sortKey, isAscendingOrder);
+        dataList =
+            Utils.sortMapObjectsByKey(dataList, sortKey, isAscendingOrder);
       }
     }
 
     templatedChildren =
         buildWidgetsFromTemplate(context, dataList, widget.itemTemplate!);
     setState(() {});
-  }
-
-  List<dynamic> sortMapObjectsByKey(
-      List<dynamic> dataMapObjects, String? sortKey, bool isAscendingOrder) {
-    final dataObjects = dataMapObjects.map((e) => e as Map).toList();
-    if (dataObjects.isNotEmpty && sortKey != null) {
-      // Function to recursively traverse the nested object and get the value of the key.
-      getValue(Map<dynamic, dynamic> mapObject, String key) {
-        if (mapObject.containsKey(key)) {
-          return mapObject[key];
-        }
-
-        for (final nestedKey in mapObject.keys) {
-          final nestedObject = mapObject[nestedKey];
-          if (nestedObject is Map<dynamic, dynamic>) {
-            final value = getValue(nestedObject, key);
-            if (value != null) {
-              return value;
-            }
-          }
-        }
-
-        return null;
-      }
-
-      // Comparator function to compare the values of the key in two map objects.
-      int compare(Map<dynamic, dynamic> a, Map<dynamic, dynamic> b) {
-        final aValue = getValue(a, sortKey);
-        final bValue = getValue(b, sortKey);
-
-        return isAscendingOrder
-            ? aValue.compareTo(bValue)
-            : bValue.compareTo(aValue);
-      }
-
-      // Sort the array of map objects using the comparator function.
-      dataObjects.sort(compare);
-      return dataObjects;
-    }
-    // Fallback - Returning same passed-in object to the caller
-    return dataMapObjects;
   }
 
   // Change ascending to descending and vice versa in dataColumnSort object
